+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonDataGridView.cs
@@ -225,6 +225,16 @@ namespace Krypton.Toolkit
         #endregion
 
         #region Public New
+        /// <inheritdoc/>
+        [Category(@"Behavior")]
+        [DefaultValue(true)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Visible)]
+        public new bool DoubleBuffered 
+        {
+            get => base.DoubleBuffered;
+            set => base.DoubleBuffered = value;
+        }
+
         /// <summary>
         /// Gets or sets the number of columns displayed in the KryptonDataGridView.
         /// </summary>